+++ back/src/main/java/com/ll/goohaeyou/auth/application/OAuth2UserCustomService.java
@@ -1,6 +1,6 @@
 package com.ll.goohaeyou.auth.application;
 
-import com.ll.goohaeyou.member.member.domain.MemberDomainService;
+import com.ll.goohaeyou.member.member.domain.service.MemberDomainService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.oauth2.client.userinfo.DefaultOAuth2UserService;
 import org.springframework.security.oauth2.client.userinfo.OAuth2UserRequest;
