+++ back/src/main/java/com/ll/goohaeyou/auth/domain/RefreshTokenRepository.java
@@ -6,6 +6,4 @@ import java.util.Optional;
 
 public interface RefreshTokenRepository extends JpaRepository<RefreshToken, Long> {
     Optional<RefreshToken> findByUserId(Long userId);
-
-    Optional<RefreshToken> findByRefreshToken(String refreshToken);
 }
