+++ back/src/main/java/com/ll/goohaeyou/category/domain/service/JobPostCategoryDomainService.java
@@ -1,4 +1,4 @@
-package com.ll.goohaeyou.category.domain;
+package com.ll.goohaeyou.category.domain.service;
 
 import com.ll.goohaeyou.category.domain.entity.Category;
 import com.ll.goohaeyou.category.domain.entity.JobPostCategory;
