+++ back/src/main/java/com/ll/goohaeyou/global/config/SecurityConfig.java
@@ -4,10 +4,10 @@ import com.ll.goohaeyou.auth.application.CustomUserDetailsService;
 import com.ll.goohaeyou.auth.application.JwtTokenProvider;
 import com.ll.goohaeyou.auth.application.OAuth2SuccessHandler;
 import com.ll.goohaeyou.auth.application.OAuth2UserCustomService;
-import com.ll.goohaeyou.auth.domain.RefreshTokenDomainService;
+import com.ll.goohaeyou.auth.domain.service.RefreshTokenDomainService;
 import com.ll.goohaeyou.auth.infrastructure.JwtFilter;
 import com.ll.goohaeyou.auth.infrastructure.OAuth2AuthorizationRequestBasedOnCookieRepository;
-import com.ll.goohaeyou.member.member.domain.MemberDomainService;
+import com.ll.goohaeyou.member.member.domain.service.MemberDomainService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
