+++ back/src/main/java/com/ll/goohaeyou/image/domain/service/ProfileImageDomainService.java
@@ -1,4 +1,4 @@
-package com.ll.goohaeyou.image.domain;
+package com.ll.goohaeyou.image.domain.service;
 
 import com.ll.goohaeyou.global.standard.anotations.DomainService;
 import com.ll.goohaeyou.jobApplication.domain.ImageStorage;
