+++ back/src/main/java/com/ll/goohaeyou/jobPost/comment/application/CommentService.java
@@ -3,12 +3,12 @@ package com.ll.goohaeyou.jobPost.comment.application;
 import com.ll.goohaeyou.global.event.notification.CommentCreatedEvent;
 import com.ll.goohaeyou.jobPost.comment.application.dto.*;
 import com.ll.goohaeyou.jobPost.comment.domain.Comment;
-import com.ll.goohaeyou.jobPost.comment.domain.CommentDomainService;
+import com.ll.goohaeyou.jobPost.comment.domain.service.CommentDomainService;
 import com.ll.goohaeyou.jobPost.comment.domain.policy.CommentPolicy;
-import com.ll.goohaeyou.jobPost.jobPost.domain.JobPostDetailDomainService;
-import com.ll.goohaeyou.jobPost.jobPost.domain.JobPostDomainService;
+import com.ll.goohaeyou.jobPost.jobPost.domain.service.JobPostDetailDomainService;
+import com.ll.goohaeyou.jobPost.jobPost.domain.service.JobPostDomainService;
 import com.ll.goohaeyou.jobPost.jobPost.domain.entity.JobPostDetail;
-import com.ll.goohaeyou.member.member.domain.MemberDomainService;
+import com.ll.goohaeyou.member.member.domain.service.MemberDomainService;
 import com.ll.goohaeyou.member.member.domain.entity.Member;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
