+++ back/src/main/java/com/ll/goohaeyou/jobPost/jobPost/application/JobPostService.java
@@ -1,22 +1,25 @@
 package com.ll.goohaeyou.jobPost.jobPost.application;
 
-import com.ll.goohaeyou.auth.domain.UserActivityService;
-import com.ll.goohaeyou.category.domain.CategoryDomainService;
-import com.ll.goohaeyou.category.domain.JobPostCategoryDomainService;
+import com.ll.goohaeyou.auth.domain.service.UserActivityService;
+import com.ll.goohaeyou.category.domain.service.CategoryDomainService;
+import com.ll.goohaeyou.category.domain.service.JobPostCategoryDomainService;
 import com.ll.goohaeyou.category.domain.entity.Category;
 import com.ll.goohaeyou.global.event.notification.PostDeadlineEvent;
 import com.ll.goohaeyou.global.event.notification.PostDeletedEvent;
 import com.ll.goohaeyou.global.standard.base.util.PaginationUtils;
 import com.ll.goohaeyou.global.standard.base.util.Util;
 import com.ll.goohaeyou.global.standard.dto.PageDto;
-import com.ll.goohaeyou.image.domain.JobPostImageDomainService;
-import com.ll.goohaeyou.jobApplication.domain.JobApplicationDomainService;
+import com.ll.goohaeyou.image.domain.service.JobPostImageDomainService;
+import com.ll.goohaeyou.jobApplication.domain.service.JobApplicationDomainService;
 import com.ll.goohaeyou.jobPost.jobPost.application.dto.*;
-import com.ll.goohaeyou.jobPost.jobPost.domain.*;
 import com.ll.goohaeyou.jobPost.jobPost.domain.entity.JobPost;
 import com.ll.goohaeyou.jobPost.jobPost.domain.entity.JobPostDetail;
 import com.ll.goohaeyou.jobPost.jobPost.domain.policy.JobPostPolicy;
-import com.ll.goohaeyou.member.member.domain.MemberDomainService;
+import com.ll.goohaeyou.jobPost.jobPost.domain.service.EssentialDomainService;
+import com.ll.goohaeyou.jobPost.jobPost.domain.service.JobPostDetailDomainService;
+import com.ll.goohaeyou.jobPost.jobPost.domain.service.JobPostDomainService;
+import com.ll.goohaeyou.jobPost.jobPost.domain.service.WageDomainService;
+import com.ll.goohaeyou.member.member.domain.service.MemberDomainService;
 import com.ll.goohaeyou.member.member.domain.entity.Member;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
